refactor(commands): move return codes to top level
make more easily accessible elsewhere
