+++ compiler_admin/__init__.py
@@ -1,5 +1,8 @@
 from importlib.metadata import version, PackageNotFoundError
 
+RESULT_SUCCESS = 0
+RESULT_FAILURE = 1
+
 try:
     __version__ = version("compiler_admin")
 except PackageNotFoundError:
