--- compiler_admin/commands/__init__.py
@@ -1,2 +1,0 @@
-RESULT_SUCCESS = 0
-RESULT_FAILURE = 1
