+++ compiler_admin/commands/info.py
@@ -1,5 +1,4 @@
-from compiler_admin import __version__ as version
-from compiler_admin.commands import RESULT_SUCCESS, RESULT_FAILURE
+from compiler_admin import __version__ as version, RESULT_SUCCESS, RESULT_FAILURE
 from compiler_admin.services.google import CallGAMCommand, CallGYBCommand
