+++ compiler_admin/commands/offboard.py
@@ -1,7 +1,7 @@
 from argparse import Namespace
 from tempfile import NamedTemporaryFile
 
-from compiler_admin.commands import RESULT_SUCCESS, RESULT_FAILURE
+from compiler_admin import RESULT_SUCCESS, RESULT_FAILURE
 from compiler_admin.commands.delete import delete
 from compiler_admin.commands.signout import signout
 from compiler_admin.services.google import (
