+++ tests/commands/test_offboard.py
@@ -1,7 +1,7 @@
 from argparse import Namespace
 import pytest
 
-from compiler_admin.commands import RESULT_FAILURE, RESULT_SUCCESS
+from compiler_admin import RESULT_FAILURE, RESULT_SUCCESS
 from compiler_admin.commands.offboard import offboard, __name__ as MODULE
