+++ tests/commands/test_reset_password.py
@@ -1,7 +1,7 @@
 from argparse import Namespace
 import pytest
 
-from compiler_admin.commands import RESULT_FAILURE, RESULT_SUCCESS
+from compiler_admin import RESULT_FAILURE, RESULT_SUCCESS
 from compiler_admin.commands.reset_password import reset_password, __name__ as MODULE
 from compiler_admin.services.google import USER_HELLO
